Move final test util
